% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emails.R
\name{attachment_count}
\alias{attachment_count}
\title{Get the Number of Attachments in an Email}
\usage{
attachment_count(email)
}
\arguments{
\item{email}{A COMIDispatch object representing an email.}
}
\value{
An integer representing the number of attachments, or NA if not found.
}
\description{
This function returns the number of attachments in a given email object.
}
\examples{
\dontrun{
email_obj <- get_email_object()  # Assume this function retrieves an email object
attachment_count(email_obj)
}
}
