% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emails.R
\name{body}
\alias{body}
\title{Get the Body of an Email}
\usage{
body(email)
}
\arguments{
\item{email}{A COMIDispatch object representing an email.}
}
\value{
A character string representing the body, or NA if not found.
}
\description{
This function returns the body of a given email object.
}
\examples{
\dontrun{
email_obj <- get_email_object()
body(email_obj)
}
}
