% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emails.R
\name{convert_comdate}
\alias{convert_comdate}
\title{Convert COMDate to POSIXct}
\usage{
convert_comdate(comdate, tz = "UTC")
}
\arguments{
\item{comdate}{A numeric COMDate.}

\item{tz}{A character string representing the time zone. Default is "UTC".}
}
\value{
A POSIXct object representing the date and time.
}
\description{
This function converts a COMDate to a POSIXct object.
}
\examples{
\dontrun{
comdate <- get_comdate()  # Assume this function retrieves a COMDate
convert_comdate(comdate)
}
}
