% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emails.R
\name{date_received}
\alias{date_received}
\title{Get the Received Date of an Email}
\usage{
date_received(email)
}
\arguments{
\item{email}{A COMIDispatch object representing an email.}
}
\value{
A POSIXct object representing the received date, or NA if not found.
}
\description{
This function returns the received date of a given email object.
}
\examples{
\dontrun{
email_obj <- get_email_object()
date_received(email_obj)
}
}
