% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emails.R
\name{date_sent}
\alias{date_sent}
\title{Get the Sent Date of an Email}
\usage{
date_sent(email)
}
\arguments{
\item{email}{A COMIDispatch object representing an email.}
}
\value{
A POSIXct object representing the sent date, or NA if not found.
}
\description{
This function returns the sent date of a given email object.
}
\examples{
\dontrun{
email_obj <- get_email_object()
date_sent(email_obj)
}
}
