% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbox.R
\name{inbox}
\alias{inbox}
\title{Get the Inbox Folder from Outlook}
\usage{
inbox(com.object = "Outlook.Application", namespace.name = "MAPI")
}
\arguments{
\item{com.object}{A character string specifying the COM object for Outlook. Default is "Outlook.Application".}

\item{namespace.name}{A character string specifying the namespace for Outlook. Default is "MAPI".}
}
\value{
A COM object representing the Inbox folder in Outlook.
}
\description{
This function initializes a COM object for Outlook and returns the Inbox folder.
}
\examples{
\dontrun{
inbox_folder <- inbox()
}

}
