% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbox.R
\name{outlook_folder}
\alias{outlook_folder}
\title{Get any Folder from Outlook}
\usage{
outlook_folder(
  folder.name = "Inbox",
  account.name = NULL,
  com.object = "Outlook.Application",
  namespace.name = "MAPI"
)
}
\arguments{
\item{folder.name}{Name of Outlook folder. Default is "Inbox".}

\item{account.name}{Name of Outlook account as it appears in Outlook. Default is NULL, which uses user's default Outlook account.}

\item{com.object}{A character string specifying the COM object for Outlook. Default is "Outlook.Application".}

\item{namespace.name}{A character string specifying the namespace for Outlook. Default is "MAPI".}
}
\value{
A COM object representing the folder in Outlook.\
}
\description{
This function initializes a COM object for Outlook and returns any folder.
}
\examples{
\dontrun{
folder <- outlook_folder(folder.name = "Sent Items")
folder <- outlook_folder(folder.name = "Inbox", account.name = "myOtherAccount@example.com")
}

}
