% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emails.R
\name{sender}
\alias{sender}
\title{Get the Sender of an Email}
\usage{
sender(email)
}
\arguments{
\item{email}{A COMIDispatch object representing an email.}
}
\value{
A character string representing the sender's name, or NA if not found.
}
\description{
This function returns the sender's name of a given email object.
}
\examples{
\dontrun{
email_obj <- get_email_object()
sender(email_obj)
}
}
