% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emails.R
\name{sender_address}
\alias{sender_address}
\title{Get the Sender's Email Address of an Email}
\usage{
sender_address(email)
}
\arguments{
\item{email}{A COMIDispatch object representing an email.}
}
\value{
A character string representing the sender's email address, or NA if not found.
}
\description{
This function returns the sender's email address of a given email object.
}
\examples{
\dontrun{
email_obj <- get_email_object()
sender_address(email_obj)
}
}
