% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emails.R
\name{subject}
\alias{subject}
\title{Get the Subject of an Email}
\usage{
subject(email)
}
\arguments{
\item{email}{A COMIDispatch object representing an email.}
}
\value{
A character string representing the subject, or NA if not found.
}
\description{
This function returns the subject of a given email object.
}
\examples{
\dontrun{
email_obj <- get_email_object()
subject(email_obj)
}
}
