% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emails.R
\name{get_emails}
\alias{get_emails}
\title{Get Most Recent Emails from Outlook Inbox}
\usage{
get_emails(most.recent.n = NULL)
}
\arguments{
\item{most.recent.n}{An integer specifying the number of most recent emails to retrieve. Default is NULL, which retrieves all emails.}
}
\value{
A list of COM objects representing the most recent emails in the Outlook inbox.
}
\description{
This function retrieves the most recent emails from the Outlook inbox.
}
\examples{
\dontrun{
recent_emails <- get_emails(most.recent.n = 10)
}

}
