% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send}
\alias{send}
\title{Send an Email via Outlook with Optional Multiple Attachments}
\usage{
send(
  to,
  from = NULL,
  cc = NULL,
  bcc = NULL,
  subject,
  body,
  attachment_paths = NULL
)
}
\arguments{
\item{to}{A character vector of recipient email addresses for the "To" field.}

\item{from}{A character string specifying the sender's email address for the "From" field.}

\item{cc}{A character vector of recipient email addresses for the "CC" field. Default is NULL.}

\item{bcc}{A character vector of recipient email addresses for the "BCC" field. Default is NULL.}

\item{subject}{A character string specifying the email subject.}

\item{body}{A character string specifying the email body.}

\item{attachment_paths}{A character vector specifying the full file paths of the attachments. Default is NULL.}
}
\value{
A message indicating the email was sent successfully.
}
\description{
This function sends an email using Microsoft Outlook. It allows you to specify recipients in the "To", "From", "CC", and "BCC" fields, and optionally attach multiple files.
}
\examples{
\dontrun{
send(
  to = c("recipient1@example.com", "recipient2@example.com"),
  from = "sender@example.com",
  cc = c("cc1@example.com", "cc2@example.com"),
  bcc = c("bcc1@example.com", "bcc2@example.com"),
  subject = "Test Subject",
  body = "This is a test email.",
  attachment_paths = c("C:/path/to/your/file1.txt", "C:/path/to/your/file2.txt")
)
}

}
